refactor using Context + add loading state when clicking enroll btn
